@@ -29,7 +29,7 @@
 		"prettier": "^3.3.2",
 		"prettier-plugin-svelte": "^3.2.6",
 		"prettier-plugin-tailwindcss": "^0.6.14",
-		"svelte": "^5.0.0",
+		"svelte": "^5.39.9",
 		"svelte-check": "^4.0.0",
 		"tailwindcss": "^4.1.13",
 		"typescript": "^5.0.0",
